% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format.plot-class.R
\docType{class}
\name{format.plot-class}
\alias{format.plot}
\alias{format.plot-class}
\title{Saving plots and output (La)TeX, html code to embed the plot in your document}
\description{
Saving plots and output (La)TeX, html code to embed the plot in your document
}
\section{Fields}{

\describe{
\item{\code{plot}}{a plot object}

\item{\code{name}}{character, name of the plot, used to save the plot in \code{path}}

\item{\code{path}}{character, path to save the plot}

\item{\code{data}}{named list, the elements are used to replace the whisker tags in your template \code{\link{whisker.render}}}

\item{\code{device}}{function name of the device to save the plot. Saves the plot with all devices that are provided, \code{\link{Devices}}}

\item{\code{plot.options}}{parameters passed to the graphical device, \code{\link{Devices}}}

\item{\code{pdf2png}}{should Ghostscript create the png from the pdf?}

\item{\code{pdf2png.options}}{options passed to \code{\link{pdf_to_png}}}

\item{\code{embed.fonts}}{logical, should the fonts be embedded in the plot, recommended for (La)TeX}

\item{\code{gs}}{path to your Ghostscript executable}

\item{\code{base54.images}}{logical, embed images base64 encoded in output, works only in non (La)TeX output}

\item{\code{overwrite}}{logical, should existing destination files be overwritten?}

\item{\code{fontpaths}}{a character vector giving directories that Ghostscript will search for fonts}
}}
\section{Methods}{

\describe{
\item{\code{initialize(plot, name, path = "figure", data = NULL, devices = "pdf",
  plot.options = NULL, pdf2png = TRUE, pdf2png.options = NULL,
  embed.fonts = TRUE, gs = Sys.getenv("R_GSCMD"), base64.images = FALSE,
  overwrite = TRUE, fontpaths = NULL)}}{Create a new format.plot object}

\item{\code{render(type = c("tex", "html"), template = NULL)}}{Render the template with the saved plot and the object fields}

\item{\code{save()}}{Save the plot with the device in .self$device}
}}

