% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdf-to-png.R
\name{pdf_to_png}
\alias{pdf_to_png}
\title{Convert PDF to PNG}
\usage{
pdf_to_png(file, outfile, fontpaths = "", resolution = 600, options = "",
  gs = Sys.getenv("R_GSCMD"), overwrite = TRUE)
}
\arguments{
\item{file}{a character string giving the name of the original pdf file}

\item{outfile}{the name of the new png file including ".png" extension}

\item{fontpaths}{a character vector giving directories that Ghostscript will search for fonts}

\item{resolution}{target resolution of the png file, default is 600}

\item{options}{a character string containing further options to Ghostscript}
}
\description{
Converts a PDf file to PNG with Ghostscript
}

